mini-services/grizzly-telegram-bot/requirements.txt
python-telegram-bot==20.7
aiohttp>=3.8.0
